Refactor bike bodels. models.py now only contain common models
